pyproject.toml
[tool.poetry]
name = "yapistir_io"
version = "0.1.0"
description = ""
authors = ["Mertcan GÖKGÖZ <mertcan.gokgoz@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.5"
django-debug-toolbar="3.5.0"
django-environ="0.9.0"
django-safedelete = "^1.2.2"
Django="^3.2.14"
docutils="0.19"
markdown="3.3.7"
Pillow="9.2.0"
psycopg2-binary="2.9.3"
pygments-lexer-solidity="0.7.0"
pygments="2.12.0"
pymdown-extensions="9.5"
requests="2.28.1"


[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
